SKARA-2588
