+++ bots/pr/src/test/java/org/openjdk/skara/bots/pr/LabelTests.java
@@ -604,69 +604,4 @@ public class LabelTests {
             assertEquals(Set.of("rfr", "group"), new HashSet<>(pr.store().labelNames()));
         }
     }
-
-    @Test
-    void autoAdjustLabel(TestInfo testInfo) throws IOException {
-        try (var credentials = new HostCredentials(testInfo);
-             var tempFolder = new TemporaryDirectory()) {
-            var author = credentials.getHostedRepository();
-            var integrator = credentials.getHostedRepository();
-
-            var censusBuilder = credentials.getCensusBuilder()
-                    .addReviewer(integrator.forge().currentUser().id())
-                    .addCommitter(author.forge().currentUser().id());
-            var labelConfiguration = LabelConfigurationJson.builder()
-                    .addMatchers("1", List.of(Pattern.compile("cpp$")))
-                    .addMatchers("2", List.of(Pattern.compile("hpp$")))
-                    .addMatchers("3", List.of(Pattern.compile("txt$")))
-                    .addGroup("group1", List.of("1", "2"))
-                    .addExtra("extra")
-                    .build();
-            var prBot = PullRequestBot.newBuilder()
-                    .repo(integrator)
-                    .censusRepo(censusBuilder.build())
-                    .labelConfiguration(labelConfiguration)
-                    .build();
-
-            // Populate the projects repository
-            var localRepoFolder = tempFolder.path().resolve("localrepo");
-            var localRepo = CheckableRepository.init(localRepoFolder, author.repositoryType(), Path.of("test.hpp"));
-            var masterHash = localRepo.resolve("master").orElseThrow();
-            assertFalse(CheckableRepository.hasBeenEdited(localRepo));
-            localRepo.push(masterHash, author.authenticatedUrl(), "master", true);
-
-            // Make a change with a corresponding PR
-            var editHash = CheckableRepository.appendAndCommit(localRepo);
-            localRepo.push(editHash, author.authenticatedUrl(), "edit", true);
-            var pr = credentials.createPullRequest(author, "master", "edit", "123: This is a pull request");
-
-            // The bot should have applied one label automatically
-            TestBotRunner.runPeriodicItems(prBot);
-            assertEquals(Set.of("2", "rfr"), new HashSet<>(pr.store().labelNames()));
-            assertLastCommentContains(pr, "The following label will be automatically applied");
-            assertLastCommentContains(pr, "`2`");
-
-            var test1Cpp = localRepo.root().resolve("test1.cpp");
-            try (var output = Files.newBufferedWriter(test1Cpp)) {
-                output.append("test");
-            }
-            localRepo.add(test1Cpp);
-            var addHash = localRepo.commit("add cpp file", "duke", "duke@openjdk.org");
-            localRepo.push(addHash, author.authenticatedUrl(), "edit", true);
-            TestBotRunner.runPeriodicItems(prBot);
-            assertEquals(Set.of("group1", "rfr"), new HashSet<>(pr.store().labelNames()));
-
-            // Simulate force-push.
-            localRepo.checkout(editHash);
-            var test1txt = localRepo.root().resolve("test1.txt");
-            try (var output = Files.newBufferedWriter(test1txt)) {
-                output.append("test");
-            }
-            localRepo.add(test1txt);
-            var forcePushHash = localRepo.commit("add txt file", "duke", "duke@openjdk.org");
-            localRepo.push(forcePushHash, author.authenticatedUrl(), "edit", true);
-            TestBotRunner.runPeriodicItems(prBot);
-            assertEquals(Set.of("group1", "rfr", "3"), new HashSet<>(pr.store().labelNames()));
-        }
-    }
 }
